Added CollapsingAvatarToolbar to library
This is where the magic happens.
Added the custom view responsible for managing avatar and title
manipulation.

Right now it's not customizable and has many harcoded values that
won't be appropiate for devices without xxhdpi density like the
Nexus 5.

Anyway, this is the starting point where I decided that my little
test could be converted to a library. Now it begins...